lib/ukschedcoop: Check for pending events before halting
The idle thread runs with interrupts enabled and interrupts could set
the `sched_have_pending_events` bit between the check and the hlt
instruction (assuming the x86 architecture). The execution would only
resume after another interrupt. Therefore, there needs to be an
additional check while the interrupts are disabled to ensure this
situation does not happen.

Signed-off-by: Marco Schlumpp <marco@unikraft.io>